lib/trace: add trace_porter tool
trace_porter is used to poll the spdk trace shm file
and store new entries from it to another specified trace file.
This could help retain the trace_entires from the overlay of
trace circular buffer

A usage sample is:
./spdk_trace_porter -s bdev_svc -p $! -f trace.tmp -q

+++ include/spdk/trace.h
@@ -102,8 +102,8 @@ struct spdk_trace_history {
 	 */
 	uint64_t			tpoint_count[SPDK_TRACE_MAX_TPOINT_ID];
 
-	/** Index to next spdk_trace_entry to fill in the circular buffer. */
-	uint32_t			next_entry;
+	/** Index to next spdk_trace_entry to fill linear. */
+	uint64_t			next_entry;
 
 	/**
 	 * Circular buffer of spdk_trace_entry structures for tracing
